Make withGradient available to RN libs.
