+++ packages/block-editor/src/components/gradients/with-gradient.js
@@ -1,0 +1,9 @@
+/**
+ * Internal dependencies
+ */
+import { __experimentalUseGradient } from './use-gradient';
+
+export const withGradient = ( WrappedComponent ) => ( props ) => {
+	const { gradientValue } = __experimentalUseGradient();
+	return <WrappedComponent { ...props } gradientValue={ gradientValue } />;
+};
